+++ custom_components/thermomaven/__init__.py
@@ -16,7 +16,7 @@ from .thermomaven_api import ThermoMavenAPI
 
 _LOGGER = logging.getLogger(__name__)
 
-PLATFORMS = [Platform.SENSOR]
+PLATFORMS = [Platform.SENSOR, Platform.CLIMATE]
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up ThermoMaven from a config entry."""